(tlg0561.tlg001.perseus-grc2.xml) Bk1 punct manual
The elision of the final vowel of some words is marked by an overstrike character. I manually have changed it to a ʼ as per other Perseus texts. Parentheses, ie round brackets are not displaying correctly, so I've manually changed them too.
I intend to go through the other books in turn, if this is a useful effort.