Correct archive.org links
According to the <sourceDesc>, the following plays are based on the third volume of Murray’s edition from 1913: Helena, Phoenissae, Orestes, Bacchae, Iphigenia in Aulis, and Rhesus.
However, the archive.org links in the files link Murray’s edition from 1909 (for the date, see the praefatio, p. vii, https://archive.org/details/euripidisfabulae03euriuoft/page/8/). The edition from 1913 is also available on archive.org: https://archive.org/details/euripidisfabulae03euri_0/page/n9/ (for the date, see the praefatio, p. vii). If the PDL files are based on the edition from 1913 and not on that from 1909, the links should be changed as suggested.